Engine/requirements.txt
pyswisseph
hijri-converter
pytest
reportlab
openpyxl
stripe
python-dotenv
fastapi
uvicorn
anthropic
httpx
hdate
